+++ CHANGELOG.md
@@ -26,7 +26,8 @@
 
 - [ibc]
   - Follow Rust guidelines naming conventions ([#689])
-
+  - Per client structure modules ([#740])
+  
 - [ibc-relayer]
   - [nothing yet]
 
@@ -67,6 +68,7 @@
 [#699]: https://github.com/informalsystems/ibc-rs/issues/699
 [#700]: https://github.com/informalsystems/ibc-rs/pull/700
 [#702]: https://github.com/informalsystems/ibc-rs/issues/702
+[#740]: https://github.com/informalsystems/ibc-rs/issues/740
 
 ## v0.1.1
 *February 17, 2021*